refactor(format): extract diff markdown renderer
