+++ docs/architecture-consolidation-plan.md
@@ -62,7 +62,7 @@ fixtures:
 | Content complexity | `crates/tokmd-analysis/src/content/complexity.rs` and `crates/tokmd-analysis/src/content/complexity/` | `tests/unit.rs` 1451; production owner modules <=187 | Scoring, nesting, and function-span helpers now live under owner modules; remaining work is mostly test split and aggregation cleanup |
 | Analysis API surface | `crates/tokmd-analysis/src/api_surface/mod.rs` and `crates/tokmd-analysis/src/api_surface/` | `mod.rs` 13; report owner 227; symbol dispatcher 56; language scanners <=68; symbol tests 385 | Keep `mod.rs` as a thin coordinator, report aggregation in `report.rs`, source scanning in language owner modules under `symbols`, and leave large integration tests under `api_surface/tests` |
 | Context packing and handoff | `crates/tokmd/src/context_pack.rs`, `crates/tokmd/src/context_pack/`, `crates/tokmd/src/commands/context.rs`, and `crates/tokmd/src/commands/handoff.rs` | `context_pack.rs` 15; selection coordinator 338 + selection tests 1215; ranking/packing owner 145; output/log submodule 206; render submodule 182; manifest submodule 180; budget parser/tests 200; context command 218; handoff command 268; handoff capability owner 113; handoff intelligence coordinator 196 + complexity owner 434; handoff output owner/tests 242 | Budget parsing, file selection, selection proof tests, ranking/packing, bundle text rendering, single-output/log writing, context bundle manifest writing, handoff capability detection, handoff intelligence construction, lightweight source-complexity estimation, and handoff artifact output writing now live in owner modules; keep context/handoff proof scoped while future handoff splits depend on fresh evidence |
-| Format core outputs | `crates/tokmd-format/src/diff.rs`, `crates/tokmd-format/src/diff/`, `crates/tokmd-format/src/export.rs`, `crates/tokmd-format/src/export/`, `crates/tokmd-format/src/summary.rs`, and `crates/tokmd-format/src/summary/` | `diff.rs` 574; diff compute owner 168; export 314; export CSV owner 55; export JSON owner 76; export JSONL owner 136; export CycloneDX owner 152; summary 557; language summary owner 95; module summary owner 58 | Diff row/total computation, CSV/JSON/JSONL/CycloneDX export rendering, and language/module summary table rendering now live in dedicated owner modules while public helpers and output behavior stay unchanged; future diff rendering or summary JSON/file-writing splits should be snapshot-backed and preserve the root `tokmd-format` re-exports |
+| Format core outputs | `crates/tokmd-format/src/diff.rs`, `crates/tokmd-format/src/diff/`, `crates/tokmd-format/src/export.rs`, `crates/tokmd-format/src/export/`, `crates/tokmd-format/src/summary.rs`, and `crates/tokmd-format/src/summary/` | `diff.rs` 313; diff compute owner 168; diff render owner 274; export 314; export CSV owner 55; export JSON owner 76; export JSONL owner 136; export CycloneDX owner 152; summary 557; language summary owner 95; module summary owner 58 | Diff row/total computation, diff Markdown rendering, CSV/JSON/JSONL/CycloneDX export rendering, and language/module summary table rendering now live in dedicated owner modules while public helpers and output behavior stay unchanged; future summary JSON/file-writing splits should be snapshot-backed and preserve the root `tokmd-format` re-exports |
 | Core context/handoff DTO contracts | `crates/tokmd-types/src/lib.rs`, `crates/tokmd-types/src/inventory.rs`, `crates/tokmd-types/src/context.rs`, and `crates/tokmd-types/src/diff.rs` | `lib.rs` 453; inventory owner 453; context owner 432; diff owner 85 | Lang/module/export/run inventory receipt DTOs, context receipts, context bundle manifests, handoff manifests, token estimation/audit DTOs, artifact/capability DTOs, and diff receipt DTOs now live in dedicated owners while root re-exports preserve the public contract |
 | Analysis DTO contracts | `crates/tokmd-analysis-types/src/lib.rs` and owner DTO modules | `lib.rs` 113; baseline owner 37 + complexity-baseline submodule 256 + complexity-section submodule 37 + determinism submodule 22 + metrics submodule 45 + file-entry submodule 23; envelope owner 24; receipt owner 42; topics owner tests 42; entropy owner tests 58; license owner tests 42; churn owner tests 50; complexity owner tests 51 + file submodule 46 + risk submodule 43 + halstead submodule 30 + maintainability submodule 25 + histogram submodule 79 + technical-debt submodule 42; effort owner 25 + estimate submodule 70 + assumptions submodule 35 + cocomo submodule 48 + model submodule 37 + confidence submodule 45 + delta submodule 56 + driver submodule 43 + size submodule 65 + results submodule 45 | Keep root receipt glue and public re-exports stable while moving remaining DTO ownership into modules |
 | Derived analysis report | `crates/tokmd-analysis/src/derived/mod.rs` and `crates/tokmd-analysis/src/derived/` | `mod.rs` 220; ratio owner 157; integrity owner 124; file metrics owner 119; language owner 88; distribution owner 80 | Keep `mod.rs` as the receipt assembly coordinator; distribution/histogram calculations, integrity hashing, density/ratio reports, top-file/offender reports, and language composition reports now live in coherent owner modules. Future derived work should start from concrete product or proof evidence rather than splitting for line count. |
